Refactor domain tests
