+++ tests/Unit/Domain/CMakeLists.txt
@@ -19,7 +19,6 @@ set(DOMAIN_TESTS
     Domain/Test_CreateInitialElement.cpp
     Domain/Test_Direction.cpp
     Domain/Test_Domain.cpp
-    Domain/Test_DomainCreators.cpp
     Domain/Test_DomainHelpers.cpp
     Domain/Test_Element.cpp
     Domain/Test_ElementId.cpp
